# openapi.yaml
...
openapi: 3.0.3
info:
  title: rucQ API
  version: 0.1.0
  description: rucQ API
servers:
  - url: http://localhost:8080
    description: 開発環境
paths:
  /api/camps:
    get:
      summary: 合宿の一覧を取得
      tags:
        - Camps
      operationId: getCamps
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Camp"
        "500":
          $ref: "#/components/responses/InternalServerError"
    post:
      summary: 合宿を作成
      tags:
        - Camps
      operationId: postCamp
      parameters:
        - $ref: "#/components/parameters/X-Forwarded-User"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PostCampRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PostCampRequest"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          $ref: "#/components/responses/Conflict"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/camps/default:
    get:
      summary: デフォルトの合宿を取得
      tags:
        - Camps
      operationId: getDefaultCamp
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Camp"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/camps/{camp_id}:
    get:
      summary: 合宿の詳細を取得
      tags:
        - Camps
      operationId: getCamp
      parameters:
        - $ref: "#/components/parameters/CampId"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Camp"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
    put:
      summary: 合宿を更新
      tags:
        - Camps
      operationId: putCamp
      parameters:
        - $ref: "#/components/parameters/CampId"
        - $ref: "#/components/parameters/X-Forwarded-User"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PostCampRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Camp"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/events:
    get:
      summary: イベントの一覧を取得
      tags:
        - Events
      operationId: getEvents
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Event"
        "500":
          $ref: "#/components/responses/InternalServerError"
    post:
      summary: イベントを作成
      tags:
        - Events
      operationId: postEvent
      parameters:
        - $ref: "#/components/parameters/X-Forwarded-User"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PostEventRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/events/{event_id}:
    get:
      summary: イベントの詳細を取得
      tags:
        - Events
      operationId: getEvent
      parameters:
        - $ref: "#/components/parameters/EventId"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
    put:
      summary: イベントを更新
      tags:
        - Events
      operationId: putEvent
      parameters:
        - $ref: "#/components/parameters/EventId"
        - $ref: "#/components/parameters/X-Forwarded-User"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PostEventRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
    delete:
      summary: イベントを削除
      tags:
        - Events
      operationId: deleteEvent
      parameters:
        - $ref: "#/components/parameters/EventId"
        - $ref: "#/components/parameters/X-Forwarded-User"
      responses:
        "204":
          $ref: "#/components/responses/NoContent"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
...
  /api/options:
    post:
      summary: 選択肢を作成
      tags:
        - Questions
      operationId: postOption
      parameters:
        - $ref: "#/components/parameters/X-Forwarded-User"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PostOptionRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Option"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/options/{option_id}:
    put:
      summary: 選択肢を更新
      tags:
        - Questions
      operationId: putOption
      parameters:
        - $ref: "#/components/parameters/OptionId"
        - $ref: "#/components/parameters/X-Forwarded-User"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PostOptionRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Option"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
...
  /api/dm:
    post:
      summary: DMを送信
      tags:
        - DM
      operationId: postDirectMessage
      parameters:
        - $ref: "#/components/parameters/X-Forwarded-User"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PostDMRequest"
      responses:
        "204":
          $ref: "#/components/responses/NoContent"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
components:
  parameters:
    X-Forwarded-User:
      name: X-Forwarded-User
      in: header
      description: ログインしているユーザーのtraQ ID（NeoShowcaseが自動で付与）
      schema:
        type: string
...
    CampId:
      name: camp_id
      in: path
      description: 合宿ID
      required: true
      schema:
        type: integer
...
    EventId:
      name: event_id
      in: path
      description: イベントID
      required: true
      schema:
        type: integer
...
    OptionId:
      name: option_id
      in: path
      description: 選択肢ID
      required: true
      schema:
        type: integer
...
  responses:
    NoContent:
      description: No Content
    BadRequest:
      description: Bad Request
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
    NotFound:
      description: Not Found
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
    Conflict:
      description: Conflict
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
    InternalServerError:
      description: Internal Server Error
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                example: Internal server error
  schemas:
    Camp:
      type: object
      properties:
        id:
          type: integer
        display_id:
          type: string
        name:
          type: string
        is_draft:
          type: boolean
        description:
          type: string
      required:
        - id
        - display_id
        - name
        - is_draft
        - description
    PostCampRequest:
      type: object
      properties:
        display_id:
          type: string
        name:
          type: string
        is_draft:
          type: boolean
        description:
          type: string
      required:
        - display_id
        - name
        - is_draft
        - description
    Event:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        location:
          type: string
        time_start:
          type: string
          format: date-time
        time_end:
          type: string
          format: date-time
        camp_id:
          type: integer
        organizer_traq_id:
          type: string
        by_staff:
          type: boolean
        display_color:
          type: string
        participants:
          type: array
          items:
            $ref: "#/components/schemas/User"
      required:
        - id
        - name
        - description
        - location
        - time_start
        - time_end
        - camp_id
        - organizer_traq_id
        - by_staff
        - display_color
        - participants
    PostEventRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        location:
          type: string
        time_start:
          type: string
          format: date-time
        time_end:
          type: string
          format: date-time
        camp_id:
          type: integer
        create_as_staff:
          type: boolean
        display_color:
          type: string
      required:
        - name
        - description
        - location
        - time_start
        - time_end
        - camp_id
        - create_as_staff
        - display_color
    User:
      type: object
      properties:
        traq_id:
          type: string
        is_staff:
          type: boolean
      required:
        - traq_id
        - is_staff
...
    Option:
      type: object
      properties:
        id:
          type: integer
        question_id:
          type: integer
        content:
          type: string
      required:
        - id
        - question_id
        - content
    PostOptionRequest:
      type: object
      properties:
        question_id:
          type: integer
        content:
          type: string
      required:
        - question_id
        - content
...
    PostDMRequest:
      type: object
      properties:
        target_user:
          type: string
        content:
          type: string
        sendtime:
          type: string
          format: date-time
      required:
        - target_user
        - content
...